requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-login
pillow
sendgrid
flask-migrate
